release: switch to macos as host for making releases to fully automate codesigning workflow
The previous release process used the dockercross image to execute CGO
compatible cross-compilation for all platforms using Linux (docker) as
the build environment. This necessitated a multi-step workflow in order
to CodeSign the macos binaries. This is a cumbersome and error prone
workflow.

This changes switches to macOS as the build environment, implementing
cross-compilation for windows and linux binaries and a goreleaser
post-build hook to codesign the macOS binary.

The release pipeline is now fully automated from CI (currently github
actions)